😿 Reverted System.Text.Json -> Newtonsoft.Json (#35)
* :crying_cat_face: Reverted System.Text.Json -> Newtonsoft.Json
System.Text.Json doesn't work for Polymorphic types. Crazy :(

Also a deal breaker.

So reverting back to the older and slower Newtonsoft.Json.

* Ignore launchSettings.json for the test project.
Just ignore everytime the app keeps recreating the launchSettings.json in this test project. It's a weird sideeffect of having a test project but also a webproject in it.